refactor(mobile): extract DogHeroNavBar with chevron icon and owner gate
Address remaining PR #173 review items:

- Extract the inline HeroNavBar from `app/dogs/[id]/index.tsx` to its own
  `components/dogs/DogHeroNavBar.tsx` named export, matching the
  per-component file convention used by other dog/* components.
- Replace the `‹` text glyph with an `IconSymbol` chevron.backward —
  iOS gets the native SF Symbol, Android falls back to MaterialIcons
  `chevron-left` via the existing IconSymbol mapping. Adds the
  `chevron.backward` entry to `icon-symbol.tsx` MAPPING.
- Gate the Edit affordance with `isOwner` so non-owner members no longer
  see a button they can't act on (matches Delete's gating). The previous
  Stack-header implementation showed Edit unconditionally; this aligns
  the UI with the actual server-side authorization.
- Convert `hitSlop={12}` to the explicit object form
  `{ top: 12, bottom: 12, left: 12, right: 12 }` per RN docs preference.
- Tests: add `getAllByText('Edit').toHaveLength(1)` regression guard to
  catch any future stray Edit duplicate, and a non-owner test that
  asserts Edit is hidden for member-role users.
