+++ apps/mobile/components/ui/icon-symbol.tsx
@@ -18,6 +18,7 @@ const MAPPING = {
   'paperplane.fill': 'send',
   'chevron.left.forwardslash.chevron.right': 'code',
   'chevron.right': 'chevron-right',
+  'chevron.backward': 'chevron-left',
   globe: 'language',
   ruler: 'straighten',
   'bell.fill': 'notifications',
